feat:add authenticate on delete expense

+++ app.js
@@ -15,10 +15,12 @@ const expenseDesc = document.querySelector("#expenseDesc");
 const output = document.querySelector(".output");
 const form = document.querySelector(".inputDiv");
 let currentItem = {};
+let expenseList = [];
 
 let expenseAmount = 0,
   expenseDescription = null;
-totalAmount = 0;
+let amtToDelete = 0;
+let totalAmount = 0;
 const inputHandler = () => {
   expenseAmt.value && expenseDesc.value
     ? (btn.disabled = false)
@@ -29,13 +31,11 @@ const btnHandler = (e) => {
   expenseAmount = expenseAmt.value;
   expenseDescription = expenseDesc.value;
 
-  totalAmount += parseInt(expenseAmount);
-  result.textContent = `Total:${totalAmount}`;
-
   currentItem.itemAmt = expenseAmount;
   currentItem.itemDesc = expenseDescription;
   currentItem.itemDate = new Date();
 
+  // result.textContent = `Total:${totalAmount}`;
   addItemDb(currentItem);
 
   expenseAmt.value = "";
@@ -57,26 +57,34 @@ function addItemDb(itemObj) {
     createdAt: d,
   });
 }
-let mainDiv;
+let mainDiv,count=0;
 function createItem(itemObj) {
+  console.log(count)
   let id = itemObj.doc.data().id;
   mainDiv = document.createElement("div");
-  mainDiv.innerHTML = `<div id="${id}" class="expenseItem">
+  mainDiv.innerHTML =`<div id="${id}" class="expenseItem">
   <div  style="display:flex;flex-direction:column;align-items:flex-start">
   <p>${itemObj.doc.data().description}</p>
   <small>${getDate(itemObj.doc.data().createdAt)}</small>
   </div>
   <div>
-  <small>${
+  <small id="amt" >${
     itemObj.doc.data().itemAmount
   } <i class="fas fa-rupee-sign"></i></small>
-  <button onclick="deleteItem(${id})"
+  <button id="delBtn" onclick="deleteItem(${id})"
   style="background:transparent;margin:0;padding:0"><i id="delete"
    class="far fa-trash-alt"></i></button>
   </div>
   </div>`;
   output.appendChild(mainDiv);
 }
+
+// function calculateTotal() {
+//   // console.log(div);
+//   // let element = div.childNodes;
+//   // amtToDelete = element[3].innerText;
+// }
+
 getData();
 function getData() {
   firebase
@@ -88,28 +96,42 @@ function getData() {
       // console.log(changes[0].doc.data().description);
       changes.forEach((item) => {
         if (item.type == "added") {
+          count++;
           createItem(item);
+          // totalAmount += Number(item.doc.data().itemAmount);
+          // console.log(totalAmount)
         } else if (item.type == "removed") {
+          count--;
           let div = document.getElementById(item.doc.data().id);
+          // calculateTotal(div);
+          // totalAmount += item.doc.data().itemAmount;
           div.remove();
+          console.log(count)
         }
       });
     });
 }
 
-function deleteItem(itemId, itemDate) {
-  console.log(itemId);
-  console.log(itemDate);
-  let query = firebase
-    .firestore()
-    .collection("expenses")
-    .where("id", "==", itemId);
-  query.get().then((snapshot) => {
-    snapshot.forEach((doc) => {
-      doc.ref.delete();
+function deleteItem(itemId) {
+  let isDelete = window.confirm(
+    `Are you sure you want to delete this expense?`
+  );
+  if (isDelete) {
+    let query = firebase
+      .firestore()
+      .collection("expenses")
+      .where("id", "==", itemId);
+    query.get().then((snapshot) => {
+      snapshot.forEach((doc) => {
+        doc.ref.delete();
+      });
     });
-  });
+  }
 }
+// let del = document.querySelector("#delBtn");
+// del.addEventListener("click", () => {
+//   totalAmount -= amtToDelete;
+// });
 form.addEventListener("submit", (e) => btnHandler(e));
 expenseAmt.addEventListener("input", inputHandler);
 expenseDesc.addEventListener("input", inputHandler);
